backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["cristian-rincon <cristian.o.rincon.b@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.2"
uvicorn = "^0.29.0"
sqlalchemy = "^2.0.29"
python-dotenv = "^1.0.1"
mysql-connector-python = "^8.4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
black = "^24.4.2"
isort = "^5.13.2"
autoflake8 = "^0.4.1"
autoflake = "^2.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
